Move collaborators into its own file
