+++ KrakenVault/Vault Module/LocalVaultLoader+Publisher.swift
@@ -1,0 +1,16 @@
+//
+//  Copyright © 2022 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Combine
+import KrakenVaultCore
+
+extension LocalVaultLoader {
+    func publisher() -> AnyPublisher<[VaultItem], Error> {
+        Deferred {
+            Future { [weak self] completion in
+                self?.load(completion: completion)
+            }
+        }.eraseToAnyPublisher()
+    }
+}
