+++ KrakenVault/Vault Module/PasswordVaultAction.swift
@@ -1,0 +1,10 @@
+//
+//  Copyright © 2022 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import KrakenVaultCore
+
+enum PasswordVaultAction {
+    case loadVault
+    case loadedVault([VaultItem])
+}
