+++ KrakenVault/Vault Module/PasswordVaultState.swift
@@ -1,0 +1,9 @@
+//
+//  Copyright © 2022 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import KrakenVaultCore
+
+struct PasswordVaultState {
+    var vaultItems: [VaultItem] = []
+}
