+++ KrakenVault/Vault Module/VaultView.swift
@@ -8,53 +8,6 @@ import KrakenVaultCore
 import PowerfulCombine
 import SwiftUI
 
-extension LocalVaultLoader {
-    func publisher() -> AnyPublisher<[VaultItem], Error> {
-        Deferred {
-            Future { [weak self] completion in
-                self?.load(completion: completion)
-            }
-        }.eraseToAnyPublisher()
-    }
-}
-
-struct PasswordVaultState {
-    var vaultItems: [VaultItem] = []
-}
-
-enum PasswordVaultAction {
-    case loadVault
-    case loadedVault([VaultItem])
-}
-
-typealias VaultEnvironment = (loader: LocalVaultLoader, scheduler: AnyDispatchQueueScheduler)
-
-func vaultReducer(state: inout PasswordVaultState, action: PasswordVaultAction, environment: VaultEnvironment) -> [Effect<PasswordVaultAction>] {
-    switch action {
-    case .loadVault:
-        return [
-            environment.loader.publisher()
-                .receive(on: environment.scheduler)
-                .replaceError(with: [])
-                .eraseToEffect()
-                .flatMap { items in
-                    Effect<PasswordVaultAction>.sync {
-                        .loadedVault(items)
-                    }
-                }.eraseToEffect(),
-        ]
-    case let .loadedVault(items):
-        state.vaultItems = items
-        return []
-    }
-}
-
-extension VaultItem: Identifiable {
-    public var id: String {
-        name + url.absoluteString
-    }
-}
-
 struct VaultView: View {
     @ObservedObject var store: Store<PasswordVaultState, PasswordVaultAction>
     @State var query: String = ""
@@ -79,6 +32,12 @@ struct VaultView: View {
         }
         .navigationViewStyle(StackNavigationViewStyle())
         .onAppear { store.send(.loadVault) }
+    }
+}
+
+extension VaultItem: Identifiable {
+    public var id: String {
+        name + url.absoluteString
     }
 }
